+++ extensions/data-plane/data-plane-cloud-http/src/main/java/org/eclipse/dataspaceconnector/dataplane/cloud/http/PresignedHttpDataPlaneHttpExtension.java
@@ -9,14 +9,16 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Siemens AG - changes to make it compatible with AWS S3 presigned URL for upload
  *
  */
 
 package org.eclipse.dataspaceconnector.dataplane.cloud.http;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import net.jodah.failsafe.RetryPolicy;
 import okhttp3.OkHttpClient;
-import org.eclipse.dataspaceconnector.dataplane.cloud.http.pipeline.CloudHttpDataSinkFactory;
+import org.eclipse.dataspaceconnector.dataplane.cloud.http.pipeline.PresignedHttpDataSinkFactory;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.DataTransferExecutorServiceContainer;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.PipelineService;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
@@ -27,7 +29,7 @@ import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 /**
  * Provides support for reading data from an HTTP endpoint and sending data to an HTTP endpoint.
  */
-public class CloudDataPlaneHttpExtension implements ServiceExtension {
+public class PresignedHttpDataPlaneHttpExtension implements ServiceExtension {
 
     @Inject
     private OkHttpClient httpClient;
@@ -54,7 +56,7 @@ public class CloudDataPlaneHttpExtension implements ServiceExtension {
     public void initialize(ServiceExtensionContext context) {
         var monitor = context.getMonitor();
 
-        var sinkFactory = new CloudHttpDataSinkFactory(httpClient, executorContainer.getExecutorService(),  monitor);
+        var sinkFactory = new PresignedHttpDataSinkFactory(httpClient, new ObjectMapper(), executorContainer.getExecutorService(),  monitor);
         pipelineService.registerFactory(sinkFactory);
     }
 }