+++ extensions/data-plane/data-plane-cloud-http/src/main/java/org/eclipse/dataspaceconnector/dataplane/cloud/http/pipeline/PresignedHttpDataAddressSchema.java
@@ -9,14 +9,17 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Siemens AG - changes to make it compatible with AWS S3 presigned URL for upload
  *
  */
 
 package org.eclipse.dataspaceconnector.dataplane.cloud.http.pipeline;
 
-public interface CloudHttpDataAddressSchema {
+public final class PresignedHttpDataAddressSchema {
     /**
      * The HTTP transfer type.
      */
-    String TYPE = "CloudHttpData";
+    public static final String TYPE = "PresignedHttpData";
+
+    private PresignedHttpDataAddressSchema() {}
 }