+++ extensions/data-plane/data-plane-http/src/main/java/org/eclipse/dataspaceconnector/dataplane/http/pipeline/HttpDataSinkFactory.java
@@ -9,12 +9,16 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Siemens AG - changes to make it compatible with AWS S3, Azure blob and AWS China S3 presigned URL for upload
  *
  */
 
 package org.eclipse.dataspaceconnector.dataplane.http.pipeline;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import okhttp3.OkHttpClient;
+import org.eclipse.dataspaceconnector.common.string.StringUtils;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.DataSink;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.DataSinkFactory;
 import org.eclipse.dataspaceconnector.spi.EdcException;
@@ -24,13 +28,18 @@ import org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchem
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.DataFlowRequest;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.ExecutorService;
 
 import static org.eclipse.dataspaceconnector.spi.result.Result.failure;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.ADDITIONAL_HEADERS;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_CODE;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_KEY;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.ENDPOINT;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.HTTP_VERB;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.TYPE;
+import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.USE_PART_NAME;
 
 /**
  * Instantiates {@link HttpDataSink}s for requests whose source data type is {@link HttpDataAddressSchema#TYPE}.
@@ -40,12 +49,14 @@ public class HttpDataSinkFactory implements DataSinkFactory {
     private final ExecutorService executorService;
     private final int partitionSize;
     private final Monitor monitor;
+    private final ObjectMapper mapper;
 
-    public HttpDataSinkFactory(OkHttpClient httpClient, ExecutorService executorService, int partitionSize, Monitor monitor) {
+    public HttpDataSinkFactory(OkHttpClient httpClient, ObjectMapper mapper, ExecutorService executorService, int partitionSize, Monitor monitor) {
         this.httpClient = httpClient;
         this.executorService = executorService;
         this.partitionSize = partitionSize;
         this.monitor = monitor;
+        this.mapper = mapper;
     }
 
     @Override
@@ -72,6 +83,9 @@ public class HttpDataSinkFactory implements DataSinkFactory {
         }
         var authKey = dataAddress.getProperty(AUTHENTICATION_KEY);
         var authCode = dataAddress.getProperty(AUTHENTICATION_CODE);
+        var httpVerb = dataAddress.getProperty(HTTP_VERB, "POST");
+        var usePartName = Boolean.parseBoolean(dataAddress.getProperty(USE_PART_NAME, Boolean.TRUE.toString()));
+        var additionalHeaders = dataAddress.getProperty(ADDITIONAL_HEADERS, "{}");
 
         return HttpDataSink.Builder.newInstance()
                 .endpoint(endpoint)
@@ -79,9 +93,22 @@ public class HttpDataSinkFactory implements DataSinkFactory {
                 .partitionSize(partitionSize)
                 .authKey(authKey)
                 .authCode(authCode)
+                .usePartName(usePartName)
+                .httpVerb(httpVerb)
+                .additionalHeaders(convertAdditionalHeaders(additionalHeaders))
                 .httpClient(httpClient)
                 .executorService(executorService)
                 .monitor(monitor)
                 .build();
+    }
+
+    private Map<String, String> convertAdditionalHeaders(String additionalHeaders) {
+        try {
+            return mapper.readValue(StringUtils.isNullOrBlank(additionalHeaders) ? "" : additionalHeaders, Map.class);
+        } catch (JsonProcessingException e) {
+            monitor.severe("Failed to set additional headers from " + additionalHeaders, e);
+        }
+
+        return new HashMap<>();
     }
 }